+++ idp_engine/Parse.py
@@ -39,7 +39,7 @@ from typing import Dict, List, Union, Optional
 
 from .Assignments import Assignments
 from .Expression import (ASTNode, Constructor, Accessor, Symbol, SymbolExpr,
-                         IfExpr, AQuantification, Quantee,
+                         IfExpr, AQuantification, Domain, Quantee,
                          ARImplication, AEquivalence,
                          AImplication, ADisjunction, AConjunction,
                          AComparison, ASumMinus, AMultDiv, APower, AUnary,
@@ -285,19 +285,19 @@ class Vocabulary(ASTNode):
                 name=CONCEPT,
                 constructors=[]),
             SymbolDeclaration(annotations='', name=Symbol(name=RELEVANT),
-                                sorts=[], out=Symbol(name=BOOL)),
+                                sorts=[], out=Domain(name=BOOL)),
             SymbolDeclaration(annotations='', name=Symbol(name=ARITY),
-                                sorts=[Symbol(name=CONCEPT)],
-                                out=Symbol(name=INT)),
+                                sorts=[Domain(name=CONCEPT)],
+                                out=Domain(name=INT)),
             SymbolDeclaration(annotations='', name=Symbol(name=ABS),
-                                sorts=[Symbol(name=INT)],
-                                out=Symbol(name=INT)),
+                                sorts=[Domain(name=INT)],
+                                out=Domain(name=INT)),
             SymbolDeclaration(annotations='', name=Symbol(name=INPUT_DOMAIN),
-                                sorts=[Symbol(name=CONCEPT), Symbol(name=INT)],
-                                out=Symbol(name=CONCEPT)),
+                                sorts=[Domain(name=CONCEPT), Domain(name=INT)],
+                                out=Domain(name=CONCEPT)),
             SymbolDeclaration(annotations='', name=Symbol(name=OUTPUT_DOMAIN),
-                                sorts=[Symbol(name=CONCEPT)],
-                                out=Symbol(name=CONCEPT))
+                                sorts=[Domain(name=CONCEPT)],
+                                out=Domain(name=CONCEPT))
             ] + self.declarations
 
     def __str__(self):
@@ -419,7 +419,7 @@ class SymbolDeclaration(ASTNode):
 
         type (string): name of the Z3 type of an instance of the symbol
 
-        domain (List): the list of possible tuples of arguments
+        in_domain (List): the list of possible tuples of arguments
 
         instances (Dict[string, Expression]):
             a mapping from the code of a symbol applied to a tuple of
@@ -463,7 +463,7 @@ class SymbolDeclaration(ASTNode):
         self.optimizable: bool = True
 
         self.type = None  # a string
-        self.domain = None  # all possible arguments
+        self.in_domain = None  # all possible arguments
         self.range = None  # all possible values
         self.instances = None  # {string: AppliedSymbol} not starting with '_'
         self.block: Optional[Block] = None  # vocabulary where it is declared
@@ -631,7 +631,7 @@ class Definition(ASTNode):
                 continue
             symbdec = SymbolDeclaration.make(
                 "_"+str(self.id)+"lvl_"+key.name,
-                key.arity, key.sorts, Symbol(name=REAL))
+                key.arity, key.sorts, Domain(name=REAL))
             self.level_symbols[key] = Symbol(name=symbdec.name)
             self.level_symbols[key].decl = symbdec
 
@@ -1118,7 +1118,7 @@ idpparser = metamodel_from_file(dslFile, memoization=True,
                                 classes=[IDP, Annotations,
 
                                          Vocabulary, Extern,
-                                         TypeDeclaration, Accessor,
+                                         TypeDeclaration, Accessor, Domain,
                                          SymbolDeclaration, Symbol,
                                          SymbolExpr,